refactor: Change Coordinate from class to struct
Change the `Coordinate` class to a struct. It is now readonly as well, meaning that the coordinates cannot be changed after initialization. One consequence is that the longitudes of polar points are not automatically set to 0; we'll see later whether this really matters.

Change several methods (`IsCloseTo`, `NorthPole/SouthPole/Origin`, `Antipode`) to properties.

Include new tests; revise old tests to the new format; all tests pass.