refactor: Change distribution of utility functions
Change the Angle static class to AngleUtilities and move the trigonometric functions from Utilities to it. Rename Utilities to CloseToUtilities. Add to AngleUtilities functions to normalize angles according to latitude, longitude, and azimuth conventions.

Most other changes are related to making the above changes work or a simple reordering of methods or rewriting of comments. However, I also make Coordinate and GreatCircle into `readonly record struct` types and remove the backing fields on their properties.

**Breaking Change**: The GreatCircle class no longer checks for polar coordinates or meridional paths. The calculations will fail for these. I also commented out the `DisplaceToLongitude` method for now since this cannot be extended to meridional paths.

**TODO**: Check which calculations need to be adjusted for meridional paths.

**TODO**: Adapt test methods to work with these changes.